working configuration compatibility between DotNetty & Helios
