+++ src/core/Akka.Remote.Tests/Akka.Remote.Tests.csproj
@@ -182,6 +182,10 @@
       <Project>{7dbd5c17-5e9d-40c4-9201-d092751532a7}</Project>
       <Name>Akka.TestKit.Xunit2</Name>
     </ProjectReference>
+    <ProjectReference Include="..\..\contrib\transports\Akka.Remote.Transport.Helios\Akka.Remote.Transport.Helios.csproj">
+      <Project>{29feaabc-e326-450a-9008-b5fcecf0115f}</Project>
+      <Name>Akka.Remote.Transport.Helios</Name>
+    </ProjectReference>
     <ProjectReference Include="..\Akka.Remote\Akka.Remote.csproj">
       <Project>{EA4FF8FD-7C53-49C8-B9AA-02E458B3E6A7}</Project>
       <Name>Akka.Remote</Name>
