+++ src/core/Akka.Remote/Configuration/Remote.conf
@@ -310,6 +310,9 @@ akka {
       trttl = "Akka.Remote.Transport.ThrottlerProvider,Akka.Remote"
     }
 
+	# necessary to keep backwards compatibility
+	helios.tcp.transport-class = "Akka.Remote.Transport.Helios.HeliosTcpTransport, Akka.Remote.Transport.Helios"
+
     ### Default configuration for the DotNetty based transport drivers
 
     dot-netty.tcp {
@@ -327,6 +330,11 @@ akka {
       applied-adapters = []
 
       transport-protocol = tcp
+
+	  # Byte order used for network communication. Event thou DotNetty is big-endian
+	  # by default, we need to switch it back to little endian in order to support
+	  # backward compatibility with Helios.
+	  byte-order = "little-endian"
 
       # The default remote server port clients should connect to.
       # Default is 2552 (AKKA), use 0 if you want a random available port
